refactor: Migrate to @toplocs/plugin-sdk for development environment
- Replace custom dev folder with plugin-sdk approach
- Add @toplocs/plugin-sdk dependency
- Create index.ts using createPluginDevelopmentEnvironment
- Update index.html to use new entry point
- Remove @dev alias from vite config
- Add index.ts to tsconfig.node.json
- Update tailwind config to include plugin-sdk styles
- Fix import paths for MainComponent -> Main.vue

This aligns wiki-plugin with the modern approach used in link-plugin,
reducing complexity and ensuring consistency across all plugins.
