+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="plugin-dev"></div>
-    <script type="module" src="/dev/main.ts"></script>
+    <script type="module" src="/index.ts"></script>
   </body>
 </html>
