+++ package.json
@@ -20,6 +20,7 @@
     "@tiptap/extension-text-align": "^2.9.1",
     "@tiptap/starter-kit": "^2.9.1",
     "@tiptap/vue-3": "^2.9.1",
+    "@toplocs/plugin-sdk": "^1.0.0",
     "axios": "^1.7.7",
     "gun": "^0.2020.1240",
     "vite-plugin-top-level-await": "^1.4.4",
