+++ src/views/MainView.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script setup lang="ts">
-import MainComponent from '@/components/MainComponent.vue';
+import MainComponent from '@/components/Main.vue';
 import { wikiProvider } from '@/composables/wikiProvider';
 
 const props = defineProps({
