+++ tsconfig.node.json
@@ -5,7 +5,8 @@
     "vitest.config.*",
     "cypress.config.*",
     "nightwatch.conf.*",
-    "playwright.config.*"
+    "playwright.config.*",
+    "index.ts"
   ],
   "compilerOptions": {
     "composite": true,
